feat: Refactor Stock management to use productId instead of id, update data types for price and quantity

+++ src/main/java/com/stock/infrastructure/adapters/input/rest/dto/response/StockDtoResponse.java
@@ -1,5 +1,7 @@
 package com.stock.infrastructure.adapters.input.rest.dto.response;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -13,9 +15,9 @@ public class StockDtoResponse {
 
     private Long productId;
 
-    private Integer quantity;
+    private int quantity;
 
-    private Double price;
+    private BigDecimal price;
 
     private boolean status;   
 }
